Add character limit in sound comments
Works both with new and old ui

https://github.com/MTG/freesound/issues/1615
