--- freesound/static/bw-frontend/styles/molecules/comments.scss
@@ -1,10 +1,0 @@
-.bw-comments_textarea {
-  display: block;
-  width: 100%;
-  padding: 20px 24px 5px;
-  resize: none;
-  border-radius: 5px;
-  background-color: $background-textarea;
-  border: none;
-  font-size: 16px;
-}
