+++ freesound/static/bw-frontend/styles/molecules/index.scss
@@ -2,7 +2,6 @@
 @import 'navbar';
 @import 'player';
 @import 'modal';
-@import 'comments';
 @import 'pagination';
 @import 'toast';
 @import 'formErrors';
